pyproject.toml
[project]
name = "project"
version = "0.1.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "python-dotenv==1.0.0",
    "aiosqlite>=0.21.0",
]

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"
